feat: asyn generate report integration whit celery
